Automatcally remove obsolete power plants.
This step occurs when the user clicks the end of round button. It finds
the highest score any player has, and removes all plants from the market
which have an equal or lower number.

Technically, this should happen immediately when that score is achieved.
But for now, we don't know when a player's turn is over, which makes
that much harder. Rather than guess, it's better to play it more
conservative while still (mostly) preserving the intent.